+++ src/rules/function-comma-newline-after/__tests__/index.js
@@ -24,11 +24,15 @@ testRule(rule, {
   }, {
     code: "a { background: linear-gradient(45deg,\n rgba(0,\n 0,\n 0,\n 1),\n red); }",
   }, {
-    code: "$map: (key: value, key2: value2)",
-    description: "Sass map ignored",
-  }, {
-    code: "$list: (value, value2)",
-    description: "Sass list ignored",
+    code: `
+      a {
+        transform: translate(
+          1px, /* comment */
+          1px
+        );
+      }
+    `,
+    description: "eol comments",
   } ],
 
   reject: [ {
@@ -71,6 +75,40 @@ testRule(rule, {
     message: messages.expectedAfter(),
     line: 2,
     column: 4,
+  }, {
+    code: `
+      a {
+        transform: translate(
+          1px,  /* comment (with trailing space) */ 
+          1px
+        );
+      }
+    `,
+    description: "eol comments",
+  } ],
+})
+
+testRule(rule, {
+  ruleName,
+  config: ["always"],
+  syntax: "scss",
+
+  accept: [ {
+    code: "$map: (key: value, key2: value2)",
+    description: "sass map ignored",
+  }, {
+    code: "$list: (value, value2)",
+    description: "sass list ignored",
+  }, {
+    code: `
+      a {
+        transform: translate(
+          1px, // line comment   
+          1px
+        );
+      }
+    `,
+    description: "eol comments",
   } ],
 })
 
@@ -122,8 +160,15 @@ testRule(rule, {
         }
     `,
   }, {
-    code: "$map: (key: value\n, key2: value2)",
-    description: "SCSS map",
+    code: `
+      a {
+        transform: translate(
+          1px, /* comment */
+          1px
+        );
+      }
+    `,
+    description: "eol comments",
   } ],
 
   reject: [ {
@@ -151,6 +196,37 @@ testRule(rule, {
     message: messages.expectedAfterMultiLine(),
     line: 2,
     column: 4,
+  }, {
+    code: `
+      a {
+        transform: translate(
+          1px, /* comment (with trailing space) */ 
+          1px
+        );
+      }
+    `,
+    description: "eol comments",
+  } ],
+})
+
+testRule(rule, {
+  ruleName,
+  config: ["always-multi-line"],
+  syntax: "scss",
+
+  accept: [ {
+    code: "$map: (key: value\n, key2: value2)",
+    description: "sass map",
+  }, {
+    code: `
+      a {
+        transform: translate(
+          1px, // line comment   
+          1px
+        );
+      }
+    `,
+    description: "eol comments",
   } ],
 })
 
@@ -180,8 +256,12 @@ testRule(rule, {
   }, {
     code: "a { background: linear-gradient(45deg\n,rgba(0, 0, 0, 1),red); }",
   }, {
-    code: "$map: (key: value,\nkey2: value2)",
-    description: "SCSS map",
+    code: `
+      a {
+        transform: translate(1px, 1px); /* comment */  
+      }
+    `,
+    description: "eol comments",
   } ],
 
   reject: [ {
@@ -211,3 +291,21 @@ testRule(rule, {
     column: 18,
   } ],
 })
+
+testRule(rule, {
+  ruleName,
+  config: ["never-multi-line"],
+  syntax: "scss",
+
+  accept: [{
+    code: "$map: (key: value,\nkey2: value2)",
+    description: "sass map",
+  }, {
+    code: `
+      a {
+        transform: translate(1px, 1px); // line comment  
+      }
+    `,
+    description: "eol comments",
+  } ],
+});