+++ src/rules/function-comma-space-after/index.js
@@ -43,7 +43,12 @@ export default function (expectation) {
 
 export function functionCommaSpaceChecker({ locationChecker, root, result, checkedRuleName }) {
   root.walkDecls(decl => {
-    valueParser(decl.value).walk(valueNode => {
+    let declValue = (() => {
+      let raws = decl.raws.value;
+      return (raws && raws.raw) || decl.value;
+    })()
+
+    valueParser(declValue).walk(valueNode => {
       if (valueNode.type !== "function") { return }
 
       if (!isStandardFunction(valueNode)) { return }
@@ -57,6 +62,9 @@ export function functionCommaSpaceChecker({ locationChecker, root, result, check
         result = result.slice(valueNode.value.length + 1)
         // Remove closing paren
         result = result.slice(0, result.length - 1)
+        // 1. Remove comments including preceeding whitespace (when only succeeded by whitespace)
+        // 2. Remove all other comments, but leave adjacent whitespace intact
+        result = result.replace(/(\ *\/(\*.*\*\/(?!\S)|\/.*)|(\/(\*.*\*\/|\/.*)))/, '')
         return result
       })()